Add index to event_type and entity_id
Since most of these will already be in the LRU and index was unlikely
to be needed.

When updating the sql docs, I realized there will be manual queries
that will not have access to the LRU so we need to have an index on
these columns (https://github.com/home-assistant/home-assistant.io/pull/26591)

If they have more than the size of the LRU entities or event types
an index is useful.