Cache decode of JWT tokens
To mitigate this I have been moving more things on the frontend to
websocket, but we cannot move cameras or external API calls to the websocket

Many of the REST api calls spend more time decoding the token than
the actual processing of the request. The actual crypto is highly optimized
but the JSON parsing and decoding is not.

closes #73558