Add descending index for states last updated time
If the states table is large it can take a long time
to find the newest state when searching the an ascending
index.

Since most of our queries return data in ascending order
we still need an ascending index for the ranged queries.

To find the newest state we need a descending index.

fixes #90113

replaces and closes #90208