Bump yalexs-ble to 2.1.6
changelog: https://github.com/bdraco/yalexs-ble/compare/v2.1.5...v2.1.6

If the lock gets jammed on auto lock the a voltage drop is reported. If we try to request an update the lock can crash.  If we see the voltage drop < 3.0 or the state goes to unknown 1/6, we need to drop the connection to allow the lock to recover.