Ensure esphome subscribes to bluetooth connection free before accepting requests
There was a small race where the connection to the esphome
device was not finished yet and the scanner was registered.
We need to make sure we get the response back from the
subscribe_bluetooth_connections_free before registering
the scanner to ensure we do not premature accept connection
requests.