Support toggling debug logging for custom components
The code that discovered the logger names that would have their level
adjusted hardcoded 'homeassistant.components.{domain}' which works great
for built in components, but not for custom ones which live at
'custom_components.{domain}'

It's possible to expose additional loggers via the "loggers" key in the
manifest, but no one is really doing that, so we can leverage `pkg_path`
to populate a reasonably reliable catch-all for this usecase.

Wrapped in a set to prevent duplicates.