Require a list of entity ids when fetching history
We no longer support fetching all the history as it selects
every entity id in the states table which can overload the
system. This was only used by the old history panel and
it was too easy to pass an empty list of entity ids to the
API which would result in a significant performance hit.

All history APIs now require at least one entity_id to
be selected.

This mirrors the change in #87937