Add h264_v4l2m2m codec and profiles to HomeKit cameras
Added support for the `h264_v4l2m2m` codec to utilize, among others,
the Raspberry Pi's hardware h.264 encoder. Unfortunately, the
implementation of the Pi's encoder doesn't use the standard h.264
profile names but expects a numerical value:
  $ v4l2-ctl --list-ctrls-menu -d 11
    h264_profile 0x00990a6b (menu)   : min=0 max=4 default=4 value=4
                 0: Baseline
                 1: Constrained Baseline
                 2: Main
                 4: High
To support that, this commit also adds the option to override the
default profile names used when passed to the FFmpeg encoding process.