Change entity category of IR light setting from control to config
As pointed out by @robertklep (https://github.com/home-assistant/core/issues/91082), this entity is really a configuration setting rather than a light control. Whether or not the entity is ultimately categorized as a light or a switch -- I would also vote for switch -- I believe it should be categorized as a configuration setting rather than as a control.