Update manifest.json
People with multiple evohome thermostats use a workaround since the official component does not support that.
The workaround is creating an additional custom_compontent and coping the original files there. Then renaming "evohome" to "evohome1" in the manifest and const.py.

Unfortunately, without the version string, the custom component won't load since that's mandatory now.
To prevent people from having to figure this out the hard way, I have added it to the file.