Add disk await entity to systemmonitor
Disk await, as reported here and by iostat, is the average time for I/O
requests issued to the device to be served. This includes the time spent
by the requests in queue and the time spent servicing them. Large spikes
in processing latency can often be traced to disk writes blocked on SSDs
flushing their caches, which show up here as large await spikes.

With no argument, the average for all I/O requests will be reported,
across all block devices. For greater granularity, a specific disk can
be provided via the argument, e.g. `mmcblk0` or `nvme0n1`.