Set entities unavailable if there was a problem fetching their data
When there was a problem fetching data for a particular board, we had no
way to indicate to the sensors that there was a problem since we only
retained card count information in the coordinator data.

This change stores a richer set of data in the coordinator so the sensor
can mark itself as unavailable when appropriate.

- Use data from coordinator when setting up sensors rather than data
  stored in config entry. The config entry will soon be slimmed down to
  only contain configured board ids