Avoid http route linear search fallback when there are multiple paths
hassio_ingress registers multiple paths with the same
canonical url (without the template vars). Since we
would only see the latest one, the fast dispatcher
would fallback to the linear search for ingress.

We now save both resources in the index so we only
have to do two lookups to find it