Make trello sensor tests more comprehensive
The sensor test is more of an integration test now where it tests the
full setup of the component and asserts on states actually pulled from
the state machine (hass). It also tests updates to the sensors and
ensures errors returned from the trello client are handled properly by
marking the corresponding sensor as unavailable.

This also makes the coordinator tests obsolete, so they have been
removed.