Load homekit_controller test data using its json loader
Since devices send lots of junk data we have a special json loader
to handle the bad data since we can do nothing about it. In
#97502 we will switch the test mocking to use what we actually
do in production to load json data, but since the homekit
controller fixtures are being loaded with that loader it will
fail. We need to load them with the homekit controller
loader since its simulating data from a test device