Enumerate available states in Prometheus startup
The Prometheus integration listens for state changes in order to
populate its internal metric table, but does not attempt to initialize
that in any way. The effect of this is that any integrations initialized
before the Prometheus integration does not have their states properly
registered until a state change occurrs.

This change adds an enumeration over all the available states after
installing the state change event listener in order to initialize the
internal metric table.

Fixes #96814