Restore sensorpush state when device becomes available
uses the new restore method from #97462

These devices feed us data in a trickle of small advertisements. The previous behavior partially restored the data because we only saw the last advertisement in the buffer, which meant one of the sensors would be randomly unavailable when the device came back into range.