baf: Raise ConfigEntryNotReady when the device has a mismatched UUID
aiobafi6 0.9.0 provides support for detecting that a device's UUID does
not match the UUID provided in the Service object. This patch updates
the integration to handle the resulting exception by aborting the config
entry setup.

This resolves a situation where a configured BAF device can be assigned
a new IP address and some other network device is assigned the prior IP
address. If the new device at the IP address is not a BAF device, then
the integration would fail to setup. But if it happens to be a BAF
device, then suddently the config entry would be connected to the wrong
device (i.e. cross-linking).

See also #98784, #97913, #92955, and #98807.