Add check for correct Python version to script/setup
I thought of this since I had to edit the setup script for my Ubuntu 22.04 machine. I understand that it may not be desirable to maintain a setup script that works on every distro.

So I'm not sure if this is worth the maintenance burden (updating the Python version) or if there is a way to have the Python version updated from a central location, or if that becomes too complex.
Interested in any feedback.