nanoleaf: add device triggers for Single and Double tap events
The tap events include a panel ID, which can be discovered by listening
to the nanoleaf_event event stream in the Developer settings.

Tapping a panel (at least, a Shapes panel) with three or more fingers
will produce a tap event for that panel.

Swipe events come through with a panel_id of -1, which is automatically
handled by aionanoleaf and replaced with None.