Reduce overhead to retry config entry setup
If the user has a lot of config entries that are retrying over and over
or the network has an issue at startup, we can end up with a lot of
config entries in a retry state.

If the config entry keeps retrying over and over again we would generate
a closure and we would have to inspect the closure each time async_call_later
was called to see what kind of function it was. We now generate a HassJob
once and let it be called over and over to avoid this overhead