overkiz: Add select for partial position (garage door)
The Somfy Dexxo smart io (io:DiscreteGarageOpenerWithPartialPositionIOComponent)
supports a third position state: partial. This state is meant to be
used to ventilate the garage.

This commit adds a select entity for the position state. The
implementation is very similar to the already existing position state
for gate's with a pedestrian position.

Note: For the updates to work properly, we'd still need
https://github.com/iMicknl/ha-tahoma/issues/167 /
https://github.com/iMicknl/ha-tahoma/pull/271 to be implemented.
See https://github.com/Somfy-Developer/Somfy-TaHoma-Developer-Mode/issues/26
for discussion with Somfy.