overkiz: Add set_cover_position_and_tilt service
Somfy J4 protect motors support setting the cover position and tilt
position in one command.

The main advantage for using this command (compared to individual
commands) is that the slides will not change the orientation if the
motor is already in the correct closure position. There's also no need
to wait for the target position to be reached before modifying the
orientation. It's also helpful for automating multiple motors, because
it allows that in less commands overall.