group: Fix executing services on new style groups
Currently a group entity id in service calls is only correctly expanded
if it uses the 'group' platform (= old style groups). Trying to execute
a service for a new style group will lead to an error message:
"Referenced entities cover.livingroom are missing or not currently
available".

The code extends the expand_entity_ids function to not only check the
'group' platform but additionally also the 'group' domain. This way
it's possible to execute services on groups that belong for example to
a 'cover' group.