Update condition.py
The Zone condition is checking if there are latitude and longitude attributes for the entity, but these are not required to see if the entity is in the zone, which is stored in the entity state.

The companion app is capable of location reporting using only the zone and not the exact location, which is better for security. Either way, if the zone-only reporting is used in the companion app, all zone conditions for that entity will fail because of the missing latitude attribute, even though the entity zone is known. 

The following error is thrown every time: "Error: In 'zone' condition: error matching person.john_doe with zone.home: entity person.john_doe has no 'latitude' attribute