Save latest short term stat id for each metadata_id on each run
We currently have to do a algorithmically complex query to find the
latest short term stat id every five minutes when stats are run. To
avoid this we now save the stat id for each metadata_id in a new
table latest_statistics_short_term so we can do a very fast lookup
to find the data between runs.

This solution was arrived it after considering many other options
for months in #94435 as other solutions required per DB queries
which made the code significantly more complex