Avoid ConfigEntry lookups in hass.config_entries.async_entries for domain index
We always want the config entry object instead of the entry_id. The
index was using the entry_id which is never the value we actually want
so it had to always do additional lookups to populate the result when
the domain_index was used.

We call async_entries a few times on nearly every discovery so its a
bit cleaner to avoid all the churn of building results to abort config
flows that already match a configured device.