Provide a better model for HomeKit service entries
We currently would show if HomeKit was in accessory mode or bridge mode
in the model field for the service. This made it hard to figure out what
kind of accessory the service was providing to HomeKit. We now show the
name of the accessory class (ie Light, HomeBridge, Camera, etc) so similar
models can be differentiated