Fix HomeKit handling of unavailable state
We should not propagate STATE_UNAVAILABLE to accessories since
its already handled by the available property.

This solves the problem where an entity goes unavailable for a moment
and the user gets a door lock/garage open notify when it becomes
available again. HomeKit polls the bridge/accessory for availability so
we should not propagate it as an event since it will end up setting
the state to an unknown state.