Suggest an ISO 8601 sample in datetime.set_value
The example datetime value proposed in the developer tools for the datetime.set_value is currently not correct. It does not comply with any of the formats supported by [`parse_datetime()`](https://github.com/home-assistant/core/blob/55bf309d2f9b5bde2025fbae9f0d101b8755893b/homeassistant/util/dt.py#L183) in `homeassistant/util/dt.py`

The problem is the slashes in the date which must be replaced by hyphens.

But this function does its first parsing attempt against the ISO 8601 format. The regex looks like a fallback approach to me. And I figured proposing a value complying with the ISO standard is the best we can do to educate users as to how to use the datetime.set_value service.