Fixed migration.py
Now the tests should all complete. I believe I somehow managed to not cleanup the states in my tests previously which is why I thought it was working. Not quite sure though. Anyways trying again clean I also got failing tests and managed to figure out that some of the if statements would fail the tests if moved to a separate function. Not sure why. However by moving out the rest it still has low enough complexity so it's fine.