Save previous unique id in entity registry when it changes
Since an integration that is listening for unique id changes may
not be running when the unique id is changed for an entity,
we need to be able to see the last unique id of the entity
so any mappings can be updated.

An alternative solution would be for anything that cares about
the unique id to store a mapping of the entity_id to unique_id
and when it sees the unique_id change for the entity it would
lookup the new unique_id in the registry. I did not think this
was a good solution because it meant we would have two sources
of truth.