Tuya Energy sensor (add_ele for CZ, KG, PC) is now working properly
Created a mecanism called VirtualStates that enables to change the behavior of the _on_device_report function of init. This enables to support the Tuya add_ele state that is the delta of consumption in KWh between the last measurement and the time of the update message.

This way, the VirtualState zeroes the state when it is not really supplied through the update message payload and when it is then the value is updated like it would normally do.

The modifications in the sensor is to support the new SensorStateClass values that allow the sum of all provided states

There seems to be one remaining bug when you restart the instance then some ghost values are added (HA bug??)