Revert to allowing cache for `power_factor` initialization
This reduces the attributes read on every HA startup (device initialization) for devices that likely do not even support the `power_factor` attribute.
Previously, this attribute was never read. It's now read once to see whether or not it's supported (and if unsupported, this will not create `power_factor` entity.)
Devices that do this can still send attribute reports (without being configured to do, since it's not a reportable attribute according to ZCL spec).