Allow startup to proceed while hassio integration fetches container stats
Currently startup waits for hassio to get container stats for all
containers because we always request them on first update. For
slower installs or installs with many containers this can delay
startup by up to 90s (or more). We now request an update of the
stats only after the entities have been added to avoid delaying
startup since everything has to wait for hassio to finish before
proceeding to stage 2