✨ Support TARGET_TEMPERATURE_RANGE for water heater entities
This patch allows to set temperature_target_{high, low} with
`water_heater.set_temperature` service.

Those values were already present before this patch but could not be set
with a service.
This patch changes the default WaterHeaterEntityFeature to include
TARGET_TEMPERATURE to be fully backward compatible.

This is heavily inspired by the same behavior on ClimateEntity.

This patch opens the way for the UI to expose/update high/low temperature targets.