Add OVOS/Neon Voice Assistants integration
This change adapts the code from the Mycroft integration to allow for
Notify support for both OpenVoice OS (OVOS) and Neon.AI
Voice Assistants. They are community-led forks of the Mycroft Private
Voice Assistant and, as Mycroft has gone bankrupt, their successors.
Neon uses OVOS as its upstream core provider.

Eventually more support for these voice assistants would be added,
but for now, this enables the previous support that the Mycroft
integration had.

This continues the work of @mikejgray at
https://github.com/home-assistant/core/pull/93837