Update telegram-bot dependency 13.1 -> 20.6
The version used is very old and missing features such as sending
messages to threads (https://core.telegram.org/api/threads). To support
this we need to upgrade to at least 20.0. Might as well update to
current latest while we are here.

There don't seem to be any breaking changes that impact our usage.