Add support for sending telegram messages to topics
Topics were added to telegram groups in November 2022, and to the
telegram-bot library in version 20.0. They are a purely additive change
that is exposed by a single parameter `message_thread_id`. Not passing
this parameter will not change the behaviour from current.

This same parameter is used to send messages to threads and messages to
topics inside groups.

https://telegram.org/blog/topics-in-groups-collectible-usernames/it?setln=en#topics-in-groups

Fixes #81888
Fixes #91750