Detect and use min python version if found
When the min python version for Home Assistant changes, it can sometimes
cause unexpected breakeage for developers not using a dev container.
The source of the breakage is not always apparent. This is especially
true for users who have multiple Python versions on their system.

This patch reads the min version from the `pyproject.toml` file and then
tries to detect a `python` executable for that version (eg.
`python3.11`) and then uses that for the venv creation. If it cannot
find this, it will fall back to using `python3` as before.

Additionally, when `script/setup` is run, it will check the min python
version that was used in the venv to ensure the requirement is met. This
will help catch the cases where a user may only have an executable for
`python3` that may or may not point to a compatible verison of Python.