Refactor bluetooth scanners for better seperation of concerns
The Home Assistant specific functions have been moved into a new class
HomeAssistantRemoteScanner which is the only part of the scanner that
knows about the hass object. This will allow more of the Bluetooth
implementation to be moved to an external library to provide for a
better development experience with an ESPHome Bluetooth proxy in
the future.