caldav: Use uppercase field names in VTODO
As per RFC5545 ical property names are always uppercase, and the `caldav` library we use automatically uppercases them while producing the ICAL stream.

But this means that different casings (including the default values from the library) are treated as different `dict` keys and will be separately iterated on. This can create duplicate entries which are commonly forbidden by RFC5545.

Use only uppercase in property names to alleviate this risk.