Update habluetooth to 2.0.0
There is a breaking change to async_register_scanner which makes
it unnecessary to pass connectable. The scanner itself knows
if its connectable or not and having to pass it lead to bugs
because there could be a disagreement between the values. As
habluetooth is only used in dev, its unlikely this change is
breaking because it would have been quite difficult to create
a custom component with a scanner before