Handle unknown state in HomeKit
The HomeKit integration should not try to dispatch unknown
state to the accessory since it cannot be handled. Instead
we preserve the last known state.

We do not mark it unavailable because this would prevent the
user from trying to activate the accessory to get it back
into a working state.

fixes #105801