Add color temp support to older HomeKit devices
The (older) HAP spec says that the color temp char must not
be used for lamps which support color which means many of
the HomeKit devices on the market omit the color temp char
when they implement hue and sat. In practice since iOS 12
adding both chars works fine and iOS handles it just fine.
iOS presents the same experience if the device supports
both chars or only hue/sat so we should mirror that behavior
as users frequently ask why they cannot set a color temp
when the accessory is bound to Home Assistant when they
could do it on iOS.