Refactor config entry storage and index
Refactors config entry storage to work the same as the
entity registry and the device registry. This abstracts
all the maintaining of the index into the ConfigEntryItems
class.

We now have an index for the unique id as well since this
each discovery will have to lookup to see if the unique
id already exists. This starts to be a problem when the
a device is discovered frequently. An example case of this
is when the user has a lot of ESPHome devices and deep
sleep configured. This means devices are constantly being
rediscovered and since there was previously no index on
unique id, it meant all the config entries in the domain
had to be enumerated every time a device woke up