Support latest Linksys Smart WiFi firmware
The existing code for this integration no longer works due to ongoing updates from Linksys breaking it, mainly the addition of an authentication requirement for a necessary endpoint.
This commit changes a few things to fix support with the latest version of Linksys Smart Wi-Fi., such as:
- Updates the URL in the request body to GetDevices3 instead of GetDevices. No idea why they changed it like that, but it works now.
- Adds a "password" config option for the router admin password.
- Adds a X-JNAP-Authorization header so that the requests are now properly authenticated. Calculated by base64 encoding the router admin password.

I did this for myself because I noticed the existing integration was not working properly on my Linksys Hydra 6 (MR2000) router. I fixed it up and it now works on my router. So I figured I would try and share it so others would not have to share my trouble.