Velbus Cover: Assume state for VMBxBL modules
(Older) VMBxBL modules do not support position reporting. So there is no
way to find out if a cover/blind is closed or open. (Newer modules do
report position.)

This patch enables `assumed_state` in that case, and simply reports the
last action as state. This is correct in most scenario's, but not all:
When a user interrupts a Close action by Stop, the assumed state will be
"Closed", but the actual state is "halfway", i.e. "Open". This is only
cosmetic, since assumed_state=True implies that the Down button will
still be enabled.