Increase unifiprotect fallback interval to 20s
We currently check auth/websocket connection every 5s. If the UFP
device is doing internal work it may reject valid auth while that
works completes. We currently try 10 times before declaring auth
invalid, but we do so in a 50s period. That does not seem to be
long enough for the slower consoles to complete whatever work is
going on.