Gracefully handle optional BatteryResponse values
The BatteryResponse class in the tesla_powerwall module contains several
optional Int and Float values.  This commit fixes the powerwall
integration to handle None values gracefully instead of just blindly
attempting to do math on the variables.

This prevents errors being logged in the Home Assistant logfile if the
user has a disabled Powerwall battery in their system.