Use eager_start when creating tasks via async_run_job
This is a change in semantics for async_run_job to function
the same as how the it runs callbacks, as soon as possible.

This is an experiment to see if anything breaks as it would
be ideal to avoid more event loop scheduling if we can.

Note that this does not change async_add_job which will continue
to ensure tasks are scheduled on the loop